Fix error when using Global State Reset with nothing to reset

+++ src/main/cliapp/src/containers/profilePage/ProfileComponent.js
@@ -35,9 +35,11 @@ export const ProfileComponent = () => {
 	const personSettingsService = new PersonSettingsService();
 
 	const globalResetHandler = () =>{
-		for(let setting of localUserInfo.settings) {
-			localStorage.removeItem(setting.settingsKey);
-			personSettingsService.deleteUserSettings(setting.settingsKey);
+		if (localUserInfo.settings) {
+			for(let setting of localUserInfo.settings) {
+				localStorage.removeItem(setting.settingsKey);
+				personSettingsService.deleteUserSettings(setting.settingsKey);
+			}
 		}
 		window.location.reload();
 	};
